Fitness Function Interface
Made an Interface function and fitness function file. We should keep all fitness functions within FitnessFunctions. May move interface in there too for simplicity. Also updated GeneticImages to take the interface in its constructor to use it for calculating fitness. May need to make an error for when a bad fitness function is passed